Fetch words via Chrome Remote Debugging protocol
It doesn't work reliably. It seems that some other activities are being
marked as the most forefront activity/target, like an extension, which
then doesn't return any data.

It seems the debugging protocols is not really geared towards a browser
that is also interactively used and not only driven by the remote
debugger, because I can't find a way to fetch the current tab.

I asked the question on the official mailing list

https://groups.google.com/forum/#!topic/chrome-debugging-protocol/6jIRM6KSRcY